remove dependency on uvloop
if needed we should be using hypercorn[uvloop] at this point. uvicorn is
not used afaik so not sure it should be a dev dependency.

(I’m actually not even sure if our current configuration uses uvloop at
all?)

uvloop seems to have some issues of its own (not present in asyncio) and
the performance claim might not be real at this point in time:
https://github.com/MagicStack/uvloop/issues/566